backend/requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
chromadb==0.4.22
openai==1.12.0
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5
pypdf==4.0.1
python-dotenv==1.0.1
pydantic==2.6.0
aiofiles==23.2.1
httpx==0.26.0
boto3==1.34.0  # AWS SDK for DynamoDB
